♻️ Refactor: 슬라이더 리사이징 로직 커스텀 훅으로 분리 (#7)
