+++ src/components/SliderSection.tsx
@@ -1,6 +1,7 @@
 import type { MediaItem } from '@/types';
+import useResizeSlider from '@/hooks/useResizeSlider';
 import MediaCard from '@/components/MediaCard';
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 
 interface SliderSectionProps {
   data: MediaItem[];
@@ -9,31 +10,11 @@ interface SliderSectionProps {
 }
 
 export default function SliderSection({ data, title }: SliderSectionProps) {
-  const [currentPage, setCurrentPage] = useState(0);
-  const [cardsPerPage, setCardsPerPage] = useState(7);
+  const { currentPage, setCurrentPage, cardsPerPage } = useResizeSlider();
+
   const containerRef = useRef<HTMLDivElement>(null);
 
   const totalPages = Math.ceil(data.length / cardsPerPage);
-
-  useEffect(() => {
-    const handleResize = () => {
-      const width = window.innerWidth;
-
-      if (width < 480) setCardsPerPage(3);
-      else if (width < 600) setCardsPerPage(4);
-      else if (width < 768) setCardsPerPage(5);
-      else if (width < 1024) setCardsPerPage(6);
-      else setCardsPerPage(7);
-
-      setCurrentPage(0);
-    };
-
-    handleResize();
-
-    window.addEventListener('resize', handleResize);
-
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
 
   const handleNext = () => {
     if (currentPage < totalPages - 1) setCurrentPage(prev => prev + 1);
